% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agreementS3.R
\name{agreement}
\alias{agreement}
\title{Raw and Chance-corrected Agreement}
\usage{
agreement(x, weights = "linear", random = T, digits = 4)
}
\arguments{
\item{x}{an n by data.frame containing n observations performed by m
observers for a given attribute of interest.}

\item{weights}{a character defining the weighting scheme for ordered factors.
options are 'linear' or 'quadratic' weights. Default is 'linear'}

\item{random}{a logical defining if the probability of expected agreement should be
computed from totally random allocation model or from the marginals
(as in Cohen's Kappa). Default is TRUE (totally random).}

\item{digits}{number of decimal places (rounding)}
}
\value{
a list with the following components (ordered by relevance):
\item{agreement}{a vector with raw and chance-corrected agreement estimates}
\item{balanced}{balanced agreement as the mean value of detailed agreement}
\item{detailed}{level-specific agreement of a factor}
\item{expected}{probability of expected agreement}
\item{p.value}{significance probability of a binomial test}
\item{n}{number of valid observations/ratings}
\item{m}{number of observers/raters}
\item{k}{number of levels}
\item{probability}{a list of withprobabilities used to compute agreement}
\item{ordered}{a logical indicating if ordered factors were analysed}
\item{scheme}{weighting scheme applied}
\item{weighted}{a logical indicating if analysis is weighted}
}
\description{
Raw and Chance-corrected Agreement
}
\details{
Current implementation generalizes the application of weighting schemes
(ordered factors) to raw agreement computation and generalizes
chance-corrected agreement to an arbitrary number of raters/rating sessions.
Statistical significance is computed from a binomial test against expected
probability of agreement and not assuming no agreement (0) as the null
hypothesis.
}
\author{
David Navega
}
