% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concordanceS3.R
\name{concordance}
\alias{concordance}
\title{Kendall's W Coefficient of Concordance}
\usage{
concordance(x, pairwise = T, digits = 4)
}
\arguments{
\item{x}{an n by m matrix or data.frame containing n observations performed
by m observers for a given attribute of interest}

\item{pairwise}{a logical indicating if pairwise detailed analysis should be
performed if more than two observers are present (columns of x.).
Default is TRUE}

\item{digits}{number of decimal places (rounding)}
}
\value{
a list with the following components:
\item{n}{number of valid observations/ratings}
\item{m}{number of observers/raters}
\item{estimate}{the estimate of Kendall's W}
\item{statistic}{test statistic for the Chi-Squared distribuition}
\item{p.value}{p-value for the computed test}
\item{pairwise}{a data.frame with pairwise analysis between observers/raters}
}
\description{
Kendall's W Coefficient of Concordance
}
\note{
Current implementations computes Kendall's W from Spearman's Rho
}
\author{
David Navega
}
