% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibmdS3.R
\name{ibmd}
\alias{ibmd}
\title{Information-based Measure of Disagreement}
\usage{
ibmd(x, digits = 4)
}
\arguments{
\item{x}{a data.frame of numeric vectors containing multiple replicated
observation performed by the same rater/observer or replicated observations
performed by different observers.}

\item{digits}{number of decimal places (rounding)}
}
\value{
a list with the following components:
\item{IBMD}{a numeric vector with the value of the IBMD}
\item{n}{an integer with the number of valid observations}
\item{k}{an integer indicating the number of observers or replications}
}
\description{
Computes Information-based Measure of Disagreement for two or more raters
}
\note{
Missing values are handle by row-wise deletion.
}
\references{
T Henriques, L Antunes, J Bernardes, M Matias, D Sato, C Costa-Santos. 2013.
Information-based measure of disagreement for more than two observers:
a useful tool to compare the degree of observer disagreement. BMC Medical
Research Methodology 13:47. https://doi.org/10.1186/1471-2288-13-47

C Costa-Santos, L Antunes, A Souto, J Bernardes. 2010. Assessment of
Disagreement: A New Information-Based Approach. Annals of Epidemiology,
20(7): 555 - 561. https://doi.org/10.1016/j.annepidem.2010.02.011
}
\author{
David Navega
}
