% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEM.R
\name{TEM}
\alias{TEM}
\title{Technical Error of Measurement}
\usage{
TEM(x)
}
\arguments{
\item{x}{a data.frame of numeric vectors containing multiple replicated
observation performed by the same rater/observer or replicated observations
performed by different observers.}
}
\value{
a list with the following components:
\item{TEM}{Technical error of measurement (in the units of x)}
\item{rTEM}{Relative technical error of measurement (as percentage)}
\item{R}{Coefficient of Reliability (dimensionless)}
\item{n}{an integer with the number of valid observations}
\item{k}{an integer indicating the number of observers or replications}
}
\description{
Compute technical error of measurement and associated normalized versions
}
\note{
rTEM, Relative TEM, represents a normalization of TEM using the mean of x.
R, Coefficient of Reliability or Variation, represents a normalization of TEM
using the variance of x.

Missing values are handle by row-wise deletion.
}
\references{
SJ Ulijaszek, DA Kerr. 1999. Anthropometric measurement error and
the assessment of nutritional status. British Journal of Nutrition, 82:
165 - 177.

CJ Utermohle, SL Zegura, GM Heathcote. 1983. Multiple Observers, Humidity,
and Choice of Precision Statistics: Factors Influencing Craniometric Data
Quality. American Journal of Physical Anthropology, 61: 85 -95.
}
\author{
David Navega
}
