% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BC2002}
\alias{BC2002}
\title{Buckberry & Chamberlain (2002) Iliac Auricular Surface Aging Dataset}
\format{a data.frame with 180 rows and 9 variables:
\describe{
\item{TO}{Transverse Organization, a factor with 5 levels.}
\item{ST}{Surface Texture, a factor vector with 5 levels.}
\item{MI}{Surface Texture, a factor vector with 3 levels.}
\item{MA}{Surface Texture, a factor vector with 3 levels.}
\item{AP}{Surface Texture, a factor vector with 3 levels.}
\item{PHASE}{Surface Texture, a factor vector with 7 levels defining the modal
age-related stage of the ilaic auricular surface}
\item{SCORE}{Surface Texture, a numeric vector with the sum of ST, TO, MI,
 MA, AP raw scores}
\item{SEX}{Sex, a factor vector with 2 levels.}
\item{AGE}{Age-at-Death, a numeric vector }
}}
\source{
\url{https://doi.org/10.1002/ajpa.10130}
}
\usage{
BC2002
}
\description{
The dataset used in Buckberry and Chamberlain (2002) to develop a revised
method for the iliac auricular surface aging.
}
\section{Citation}{


Buckberry JL, Chamberlain AT. 2002. Age estimation from the auricular surface
of the ilium: A revised method. American Journal of Physical Anthropology,
119(3): 231-239
}

\keyword{datasets}
